First draft Energy Savings POLICY_TO_SDNR handler
Initial commit of a karaf feature to process a DMaaP topic sent by
Policy to SDNR as part of the Energy Savings project.  The topic name
is POLICY_TO_SDNR, and the content is a large and complex JSON
structure. Currently, this feature simply forwards the JSON string to
a server and returns the response of the server to Policy in topic
SDNR_TO_POLICY.

Change-Id: Iaaa2421d97e8ea40ebe8000b3b1b93efa7980ea0
Issue-ID: CCSDK-350
Signed-off-by: George Clapp <gc731h@att.com>