Update patch set 1
Patch Set 1: Code-Review+2

FYI - this is fine for now, but wanted to bring to your attention that ODL has announced they are removing the oAuth identity provider from OpenDaylight in their Sodium release (not the next release, but the release after).  See https://jira.opendaylight.org/browse/AAA-173 for more details.

Not a problem for our Dublin or El Alto release, but for ONAP F release, you'll need to revisit this code.

Patch-set: 1
Label: Code-Review=+2