Proposal to remove OSGi dependencies from the SDNC project
Dependencies on the OSGi frameworks and libraries are removed
by integrating the CCSDK and SDNC project with the lighty.io.
It's a toolkit that allows to use ODL services (in this case
core services and the Restconf) without the dependency
on the Karaf framework and the Blueprint DI.

In this change are created the lighty.io modules which
initialize and expose same services as the Blueprint DI in
the blueprint.xml files.

More info about the lighty.io - https://lighty.io

Signed-off-by: Samuel Kontris <samuel.kontris@pantheon.tech>
Change-Id: Ia1c96aa5bc1f0fd66fd0d56b5af50ed5ed159bd9