Update patch set 1
Patch Set 1:

The Job builder hung two times at the same location. The builder stuck in old bundles that needs anyway to be removed/replaced. So I did a further change on top of this and removed the parts. The Jobbuilder was running in the patches that are on top of this one. 
Can you approve this .. or should I abandon all the changes and start with removing the old bundles?

Patch-set: 1