Update patch set 1
Patch Set 1: Code-Review+1

The java try patter for Autoclosable does not work here, because the close for the transaction will force an immediate (and implicit) cancel request to the other cluster nodes. This could be a design problem in the interface. Close must not be called.

Patch-set: 1
CC: Gerrit User 865 <865@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
CC: Gerrit User 87 <87@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Reviewer: Gerrit User 4931 <4931@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1