Update patch set 1
Patch Set 1:

This check-in is intended to provide Sodium SR3 compatibility of the SDN-R applications. You requested this indirectly via comment to https://gerrit.onap.org/r/c/ccsdk/parent/+/109831.
So the code is mainly adjusting to Sodium and tested with the parent provided by Micha. 
Additionally there was also the action to fix the "check-style" problem that we had. Unfortunately there are a lot of changes because "tabs" are replaced by "spaces". We like to improve here to "not" getting such feedback that to many has changed.

To avoid incompatibility I had to remove to bundles. This bundles are containing yang-tools generated classes. This classes are not compatible between Neon and Sodium. For this reason I had removed the related bundle with all the generated classes. Looks as a big change. But was necessary to provide the requested Neon/Sodium compatibility of the code. This were about 150 deletions. 

This code is tested now with local Sodium builds and the "expected" parents that are provided via CCSDK-2518 and Neon builds. 
The Easticsearch version that you commented is used for a plugin that is doing JUnit tests and is not used in the running Environment. Proposal is to change this version till M1 to avoid hickups.

Patch-set: 1