Update patch set 2
Patch Set 2:

Hi, Please review and approve if everything OK. Just some info - The number of files committed are very big in number but please be aware that it is just a reorganization of the directory structure of the devicemanagers and nothing else. The base devicemanager which was till now in the "devicemanager" directory is now renamed as "devicemanager-core". 2 new directories - devicemanager-onap and devicemanager-o-ran-sc were created and the specific devicemanagers are moved to those new directories.

Regards,
Ravi

Patch-set: 2