Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (1 comment)
> 
> Hi Rafal, One comment to understand the need for this change.
> 
> Thanks
> Ravi

Hi Ravi,

There are several reasons. First of all, as you mentioned, we take into consideration, that someone can put message directly into DMaaP. We also don't want to duplicate VES colletor validation, so we left it as interface, that all consumers can have it's own way of validation. And finally this approach allows as to check values of fields in incoming messages. Let's say in our fault message there will be some wierd value in one of required fields, that will make all event as a trash and unusefull in further processing. Or let's say in our nearly comming commits about CM events we will want to know if some objects exists, if they have to according to other fields value. This allows us to stop processing and not send further or save into DB events, that are in some way corrupted from our point of view.

Best,
Rafal

Patch-set: 2