Update patch set 9
Patch Set 9: Code-Review-1

> Patch Set 7: Code-Review-1
> 
> Regarding "sdnr" parts we (highstreet) like to proceed with this change like this:
> - Not mounting this change into gerrit as it is. 
> Reason: This xlarge amount of changes in one go are to risky that things get broken. The test coverage of Junit and Integration tests in gerrit does not cover enough.
> Proposal for proceeding:
> - We synchronize this changes with our local repository where we have more test cases. (We need to do this anyways).
> - In a further step we sync the local repo back to gerrit/master.
> 
> We handle all files that begin with sdnr/...   documenting progress via the Jira-issue
> and not the files with lib/...

I noticed that some of these changes have come in through other reviews.  Is the plan to abandon this change and ingest it in pieces?  Or is it safe to merge this change?

Patch-set: 9
Reviewer: Gerrit User 87 <87@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1