Chore: Update Python to 3.12 in docs/tox.ini
To fix:
https://github.com/onap/.github/actions/runs/20059755676/job/57533545148

Removes:
sphinxcontrib.swaggerdoc

Includes a stale dependency on the "imp" module, removed from Python in 3.12.

- Not used anywhere in the documentation
- Blocking Python 3.12 compatibility
- Unmaintained since 2017

Also, sphinxcontrib-swaggerdoc===0.1.7 is referenced at the URL here:
https://git.onap.org/doc/plain/etc/upper-constraints.onap.txt?h=master

It seems this is being sourced in the tox.ini setup, which will break any
tox doc job containing this reference when we attempt to upgrade the Python
version.

Issue-ID: CIMAN-33
Change-Id: Ie9b50f4d73dc6d901ee9c1397c9061a46e89c706
Signed-off-by: Matthew Watkins <mwatkins@linuxfoundation.org>